refactor: UserInfoResponse 패키징, 네이밍 변경
