+++ src/main/java/com/mobble/mobbleserver/infrastructure/oauth/provider/userInfoResult/AppleUserInfoResult.java
@@ -1,4 +1,4 @@
-package com.mobble.mobbleserver.account.auth.oauth.dto.response;
+package com.mobble.mobbleserver.infrastructure.oauth.provider.userInfoResult;
 
 import com.auth0.jwt.interfaces.DecodedJWT;
 import com.mobble.mobbleserver.application.account.command.SocialProvider;
@@ -6,13 +6,13 @@ import com.mobble.mobbleserver.global.exception.common.DomainException;
 import com.mobble.mobbleserver.global.exception.errorCode.oAuth.OAuthErrorCode;
 
 
-public class AppleUserInfoResponse implements OAuth2UserInfo {
+public class AppleUserInfoResult implements OAuth2UserInfo {
 
     private final DecodedJWT jwt;
 
     private final SocialProvider socialProvider = SocialProvider.APPLE;
 
-    public AppleUserInfoResponse(DecodedJWT jwt) {
+    public AppleUserInfoResult(DecodedJWT jwt) {
         if (jwt == null || jwt.getSubject() == null) throw new DomainException(OAuthErrorCode.NO_USER_INFO);
         this.jwt = jwt;
     }
