+++ src/main/java/com/mobble/mobbleserver/infrastructure/oauth/provider/userInfoResult/KakaoUserInfoResult.java
@@ -1,4 +1,4 @@
-package com.mobble.mobbleserver.account.auth.oauth.dto.response;
+package com.mobble.mobbleserver.infrastructure.oauth.provider.userInfoResult;
 
 import com.mobble.mobbleserver.application.account.command.SocialProvider;
 import com.mobble.mobbleserver.global.exception.common.DomainException;
@@ -6,13 +6,13 @@ import com.mobble.mobbleserver.global.exception.errorCode.oAuth.OAuthErrorCode;
 
 import java.util.Map;
 
-public class KakaoUserInfoResponse implements OAuth2UserInfo {
+public class KakaoUserInfoResult implements OAuth2UserInfo {
 
     private final Map<String, Object> attributes;
 
     private final SocialProvider socialProvider = SocialProvider.KAKAO;
 
-    public KakaoUserInfoResponse(Map<String, Object> attributes) {
+    public KakaoUserInfoResult(Map<String, Object> attributes) {
         if (attributes == null || attributes.get("id") == null || attributes.get("kakao_account") == null) throw new DomainException(OAuthErrorCode.NO_USER_INFO);
         this.attributes = attributes;
     }
