+++ src/main/java/com/mobble/mobbleserver/infrastructure/oauth/provider/userInfoResult/NaverUserInfoResult.java
@@ -1,4 +1,4 @@
-package com.mobble.mobbleserver.account.auth.oauth.dto.response;
+package com.mobble.mobbleserver.infrastructure.oauth.provider.userInfoResult;
 
 import com.mobble.mobbleserver.application.account.command.SocialProvider;
 import com.mobble.mobbleserver.global.exception.common.DomainException;
@@ -6,13 +6,13 @@ import com.mobble.mobbleserver.global.exception.errorCode.oAuth.OAuthErrorCode;
 
 import java.util.Map;
 
-public class NaverUserInfoResponse implements OAuth2UserInfo {
+public class NaverUserInfoResult implements OAuth2UserInfo {
 
     private final Map<String, Object> response;
 
     private final SocialProvider socialProvider = SocialProvider.NAVER;
 
-    public NaverUserInfoResponse(Map<String, Object> rawResponse) {
+    public NaverUserInfoResult(Map<String, Object> rawResponse) {
         if (rawResponse == null || rawResponse.get("response") == null) throw new DomainException(OAuthErrorCode.NO_USER_INFO);
         this.response = (Map<String, Object>) rawResponse.get("response");
     }
